YUK-16874 Automatically Start and Save Gateway Simulator Settings YUKRV-10973
Removed double type from input factory and its use by the gateway simulator service because long type is sufficient.

+++ yukon-client/common/src/com/cannontech/common/rfn/simulation/service/impl/RfnGatewaySimulatorServiceImpl.java
@@ -711,7 +711,7 @@ public class RfnGatewaySimulatorServiceImpl implements RfnGatewaySimulatorServic
     public SimulatedGatewayDataSettings getGatewayDataSettings() {
         if (gatewayDataSettings == null) {
             SimulatedGatewayDataSettings settings = new SimulatedGatewayDataSettings();
-            settings.setCurrentDataStreamingLoading(yukonSimulatorSettingsDao.getDoubleValue(YukonSimulatorSettingsKey.GATEWAY_SIMULATOR_DATA_STREAMING_LOADING));
+            settings.setCurrentDataStreamingLoading(yukonSimulatorSettingsDao.getLongValue(YukonSimulatorSettingsKey.GATEWAY_SIMULATOR_DATA_STREAMING_LOADING));
             settings.setReturnGwy800Model(yukonSimulatorSettingsDao.getBooleanValue(YukonSimulatorSettingsKey.GATEWAY_SIMULATOR_RETURN_GWY800_MODEL));
             settings.setNumberOfNotReadyNodes(yukonSimulatorSettingsDao.getIntegerValue(YukonSimulatorSettingsKey.GATEWAY_SIMULATOR_NUM_NOT_READY_NODES));
             settings.setNumberOfReadyNodes(yukonSimulatorSettingsDao.getIntegerValue(YukonSimulatorSettingsKey.GATEWAY_SIMULATOR_NUM_READY_NODES));
